Regold non-affine tests
In libmesh/libmesh#2997 we greatly tightened the affine map detection
tolerance. Affine map detection is mean to be an optimization only: it
should not affect results. However, previously our detection tolerance
was too loose such that we were introducing errors roughly of order 1e-6
into results for slightly deformed meshes, e.g. we would define an
element as affine when it was slightly deformed when we should have been
defining it as non-affine. With the new tightening of affine detection
we are generating results that are more accurate (albeit a bit more
slowly!)